+++ src/TreeViewProps.ts
@@ -1,5 +1,5 @@
 import React from "react";
-import { DropIndication } from "./DropIndicator";
+import { DropIndication } from "./DropIndication";
 import { TreeViewItem } from "./TreeViewItem";
 
 export interface TreeViewProps<T extends TreeViewItem> {
